requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
asttokens==2.4.1
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
comm==0.2.2
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==3.0.11
dbus-python==1.2.16
debugpy==1.8.5
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker-pycreds==0.4.0
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.0
fasteners==0.14.1
filelock==3.15.4
fsspec==2024.6.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
graphsurgeon==0.4.6
h11==0.14.0
httplib2==0.14.0
idna==2.8
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jetson-stats==4.2.9
Jetson.GPIO==2.1.6
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.1.2
matplotlib-inline==0.1.7
monotonic==1.5
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
onnx-graphsurgeon==0.3.12
packaging==24.1
PAM==0.4.2
pandas==0.25.3
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
pure_eval==0.2.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-dbusmock==0.19
python-debian==0.1.36+ubuntu1.1
python-multipart==0.0.9
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.1.0
requests==2.32.3
requests-unixsocket==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
sentry-sdk==2.13.0
setproctitle==1.3.3
simplejson==3.16.0
six==1.14.0
smbus2==0.4.3
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.2
systemd-python==234
tensorrt==8.5.2.2
threadpoolctl==3.5.0
torch==2.1.0a0+41361538.nv23.6
torchvision==0.16.2+c6f3977
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uff==0.6.9
urllib3==2.2.2
urwid==2.0.1
uvicorn==0.30.6
wadllib==1.3.3
wandb==0.17.6
wcwidth==0.2.13
xkit==0.0.0
zipp==3.20.0
